Pre alfa of a new GUI.
Major changes:

Removed the collectionActivity class and replaced with BrowseActivity, PlaylistActivity.

Added a Global Sliding drawer to be able to control playback anywhere in the program.

Removed alot of functinality from the former playlistActivity and put it in the GlobalMediaPlayerControl instead.

Added a proper search activity with the possiblity to further filter the searches. I.e. on Artist, album, tag, song or all.

Added "tabs" to the GUI to switch between activities. To do this in a good way an activitygroup has been created to host the "tabs" and the other activities

Minor changes:

Added new icons.

Added new layouts for all newly created Activities.

Broken features:

TextFiltering does not work in any listview anymore.

It is no longer possible to spawn a custom quick search dialog. This feature has been replaced by the songSearch activity. The global quick search will be displayed if pressing the search button now.

Caching of requests to the ampache server no longer exist.

 Changes to be committed:
   (use "git reset HEAD <file>..." to unstage)

	modified:   AndroidManifest.xml
	modified:   default.properties
	new file:   res/drawable/handle.png
	new file:   res/drawable/ic_btn_speak_now.png
	new file:   res/drawable/ic_menu_forward.png
	new file:   res/drawable/ic_menu_home.png
	new file:   res/drawable/ic_menu_more.png
	new file:   res/drawable/ic_menu_play_clip.png
	new file:   res/drawable/ic_menu_revert.png
	new file:   res/drawable/ic_menu_revert_disabled.png
	new file:   res/drawable/ic_menu_shuffle.png
	new file:   res/drawable/ic_menu_shuffle_disabled.png
	new file:   res/drawable/ic_menu_sort_by_size.png
	new file:   res/drawable/ic_search_category_default.png
	new file:   res/drawable/the_shit.png
	new file:   res/layout/amdroidactivitygroup_layout.xml
	modified:   res/layout/browsable_item.xml
	new file:   res/layout/browseactivity_layout.xml
	new file:   res/layout/dashactivity_layout.xml
	new file:   res/layout/list_item_music_root.xml
	new file:   res/layout/menuview_layout.xml
	modified:   res/layout/my_media.xml
	modified:   res/layout/playlist.xml
	modified:   res/layout/playlist_item.xml
	new file:   res/layout/playlistsactivity_layout.xml
	new file:   res/layout/searchactivity_layout.xml
	new file:   res/layout/slider_layout.xml
	new file:   res/menu/dash_menu.xml
	modified:   res/menu/playlist_menu.xml
	new file:   res/values/search_criterias.xml
	modified:   res/values/strings.xml
	new file:   src/com/sound/ampache/AmdroidActivityGroup.java
	new file:   src/com/sound/ampache/BrowseActivity.java
	new file:   src/com/sound/ampache/DataHandler.java
	new file:   src/com/sound/ampache/GlobalMediaPlayerControl.java
	new file:   src/com/sound/ampache/PlaylistsActivity.java
	modified:   src/com/sound/ampache/amdroid.java
	modified:   src/com/sound/ampache/ampacheCommunicator.java
	deleted:    src/com/sound/ampache/collectionActivity.java
	modified:   src/com/sound/ampache/dashActivity.java
	modified:   src/com/sound/ampache/playlistActivity.java
	modified:   src/com/sound/ampache/songSearch.java
	modified:   src/com/sound/ampache/staticMedia.java